Cleaned up project and made it work.
I haven't tested it yet, but I at least got rid of all the errors.
