+++ CardGameSuiteApp/app/src/main/java/com/example/cardgamesuiteapp/decks/Standard.java
@@ -13,8 +13,23 @@ public class Standard extends Deck {
      * @param numPlayers         How many people are playing.
      */
     @SuppressWarnings("unchecked")
-    public Standard(boolean shuffleOnEmptyDeck, int numPlayers, int myPlayerNum, boolean multiplayer) {
+    public Standard(boolean shuffleOnEmptyDeck, int numPlayers, int myPlayerNum) {
         super(shuffleOnEmptyDeck, numPlayers, myPlayerNum);
+        Queue<Integer> deck = new LinkedList<Integer>();
+        for (int i = 1; i <= 13 * 4; i++) {
+            deck.add(i);
+        }
+        ArrayList<Integer>[] hands;
+        hands = new ArrayList[numPlayers];
+        for (int i = 0; i < hands.length; i++) {
+            hands[i] = new ArrayList<Integer>();
+        }
+        super.initializeDeckAndHands(deck, hands);
+        this.shuffleDiscardIntoDeck();
+    }
+
+    public Standard(boolean shuffleOnEmptyDeck, int numPlayers) {
+        super(shuffleOnEmptyDeck, numPlayers);
         Queue<Integer> deck = new LinkedList<Integer>();
         for (int i = 1; i <= 13 * 4; i++) {
             deck.add(i);
