+++ CardGameSuiteApp/app/src/main/java/com/example/cardgamesuiteapp/display/Card.java
@@ -1,18 +1,13 @@
 package com.example.cardgamesuiteapp.display;
 
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.graphics.Canvas;
-import android.graphics.drawable.Drawable;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.View;
 
-import com.example.cardgamesuiteapp.gameCollectionMainMenu.DisplaySettingsActivity;
 import com.example.cardgamesuiteapp.games.Fives;
 import com.example.cardgamesuiteapp.games.Solitaire;
-import com.example.cardgamesuiteapp.games.FivesGame;
-import com.example.cardgamesuiteapp.R;
 import com.example.cardgamesuiteapp.decks.Standard;
 
 public class Card extends View implements View.OnTouchListener {
