+++ CardGameSuiteApp/app/src/main/java/com/example/cardgamesuiteapp/display/CardAnimation.java
@@ -6,8 +6,7 @@ import android.view.View;
 import android.view.animation.Animation;
 import android.view.animation.TranslateAnimation;
 
-import com.example.cardgamesuiteapp.games.FivesGame;
-import com.example.cardgamesuiteapp.R;
+import com.example.cardgamesuiteapp.games.Fives;
 
 public class CardAnimation {
 
@@ -42,7 +41,7 @@ public class CardAnimation {
             public void onAnimationEnd(Animation a) {
                 cardToAnimate.animate().alpha(0f).setDuration(0);
                 if (callPostAnimation) {
-                    FivesGame.postAnimation();
+                    Fives.postAnimation();
                 }
             }
         });
