+++ CardGameSuiteApp/app/src/main/java/com/example/cardgamesuiteapp/gameCollectionMainMenu/DisplayMainPageActivity.java
@@ -9,7 +9,7 @@ import android.view.View;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.cardgamesuiteapp.austenMPStuff.MultiplayerOrSinglePlayerMenu;
-import com.example.cardgamesuiteapp.games.FivesGame;
+import com.example.cardgamesuiteapp.games.Fives;
 import com.example.cardgamesuiteapp.R;
 import com.example.cardgamesuiteapp.games.Solitaire;
 import com.example.cardgamesuiteapp.singlePlayerMenus.FivesSinglePlayerMenu;
@@ -59,7 +59,7 @@ public class DisplayMainPageActivity extends AppCompatActivity implements Serial
 
     public void clickPlayFivesButton(View view) {
         Intent intent = new Intent(this, MultiplayerOrSinglePlayerMenu.class);
-        intent.putExtra("gameClass", FivesGame.class);
+        intent.putExtra("gameClass", Fives.class);
         intent.putExtra("gameName", "Fives");
         intent.putExtra("singlePlayerClass", FivesSinglePlayerMenu.class);
         startActivity(intent);
