+++ CardGameSuiteApp/app/src/main/res/layout/fives_offline_player_menu.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".games.FivesGame">
+    tools:context=".games.Fives">
 
     <LinearLayout
         android:layout_width="match_parent"
