+++ CardGameSuiteApp/app/src/main/res/layout/fives_two_players.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".games.FivesGame">
+    tools:context=".games.Fives">
 
     <Button
         android:id="@+id/confirmButton"
